requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
psycopg2-binary>=2.9.9
Werkzeug==3.1.3
python-dotenv==1.0.1
cryptography==44.0.0
gunicorn==21.2.0
waitress==2.1.2
